+++ pkg/controller/jaeger/clusterrolebinding_test.go
@@ -63,6 +63,10 @@ func TestClusterRoleBindingsUpdate(t *testing.T) {
 	orig := rbac.ClusterRoleBinding{}
 	orig.Name = nsn.Name
 	orig.Annotations = map[string]string{"key": "value"}
+	orig.Labels = map[string]string{
+		"app.kubernetes.io/instance":   orig.Name,
+		"app.kubernetes.io/managed-by": "jaeger-operator",
+	}
 
 	objs := []runtime.Object{
 		v1.NewJaeger(nsn),
@@ -101,6 +105,10 @@ func TestClusterRoleBindingsDelete(t *testing.T) {
 
 	orig := rbac.ClusterRoleBinding{}
 	orig.Name = nsn.Name
+	orig.Labels = map[string]string{
+		"app.kubernetes.io/instance":   orig.Name,
+		"app.kubernetes.io/managed-by": "jaeger-operator",
+	}
 
 	objs := []runtime.Object{
 		v1.NewJaeger(nsn),