+++ pkg/controller/jaeger/route_test.go
@@ -72,6 +72,10 @@ func TestRoutesUpdate(t *testing.T) {
 	orig := osv1.Route{}
 	orig.Name = nsn.Name
 	orig.Annotations = map[string]string{"key": "value"}
+	orig.Labels = map[string]string{
+		"app.kubernetes.io/instance":   nsn.Name,
+		"app.kubernetes.io/managed-by": "jaeger-operator",
+	}
 
 	objs := []runtime.Object{
 		v1.NewJaeger(nsn),
@@ -114,6 +118,10 @@ func TestRoutesDelete(t *testing.T) {
 
 	orig := osv1.Route{}
 	orig.Name = nsn.Name
+	orig.Labels = map[string]string{
+		"app.kubernetes.io/instance":   nsn.Name,
+		"app.kubernetes.io/managed-by": "jaeger-operator",
+	}
 
 	objs := []runtime.Object{
 		v1.NewJaeger(nsn),