+++ test/e2e/utils.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"os/exec"
 	"runtime"
+	"strconv"
 	"strings"
 	"testing"
 	"time"
@@ -15,6 +16,7 @@ import (
 	osv1sec "github.com/openshift/api/security/v1"
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
 	"github.com/operator-framework/operator-sdk/pkg/test/e2eutil"
+	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
@@ -35,7 +37,7 @@ var (
 	timeout              = time.Minute * 2
 	storageNamespace     = os.Getenv("STORAGE_NAMESPACE")
 	kafkaNamespace       = os.Getenv("KAFKA_NAMESPACE")
-	noSetup              = os.Getenv("NO_SETUP")
+	usingOLM             = getBoolEnv("OLM", false)
 	esServerUrls         = "http://elasticsearch." + storageNamespace + ".svc:9200"
 	cassandraServiceName = "cassandra." + storageNamespace + ".svc"
 	ctx                  *framework.TestCtx
@@ -43,6 +45,17 @@ var (
 	namespace            string
 	t                    *testing.T
 )
+
+func getBoolEnv(key string, defaultValue bool) bool {
+	if value, ok := os.LookupEnv(key); ok {
+		boolValue, err := strconv.ParseBool(value)
+		if err != nil {
+			logrus.Warnf("Error [%v] received converting environment variable [%s] using [%v]", err, key, boolValue)
+		}
+		return boolValue
+	}
+	return defaultValue
+}
 
 // GetPod returns pod name
 func GetPod(namespace, namePrefix, containsImage string, kubeclient kubernetes.Interface) corev1.Pod {
@@ -71,8 +84,8 @@ func GetPod(namespace, namePrefix, containsImage string, kubeclient kubernetes.I
 
 func prepare(t *testing.T) (*framework.TestCtx, error) {
 	ctx := framework.NewTestCtx(t)
-	doSetup := len(noSetup) == 0
-	if doSetup {
+	// Install jaeger-operator unless we've installed it from OperatorHub
+	if !usingOLM {
 		err := ctx.InitializeClusterResources(&framework.CleanupOptions{TestContext: ctx, Timeout: timeout, RetryInterval: retryInterval})
 		if err != nil {
 			t.Fatalf("failed to initialize cluster resources: %v", err)
@@ -119,7 +132,7 @@ func prepare(t *testing.T) (*framework.TestCtx, error) {
 	// get global framework variables
 	f := framework.Global
 	// wait for the operator to be ready
-	if doSetup {
+	if !usingOLM {
 		err = e2eutil.WaitForDeployment(t, f.KubeClient, namespace, "jaeger-operator", 1, retryInterval, timeout)
 		if err != nil {
 			return nil, err
@@ -266,6 +279,7 @@ type services struct {
 }
 
 func createJaegerInstanceFromFile(name, filename string) *v1.Jaeger {
+	// #nosec   G204: Subprocess launching should be audited
 	cmd := exec.Command("kubectl", "create", "--namespace", namespace, "--filename", filename)
 	output, err := cmd.CombinedOutput()
 	if err != nil && !strings.Contains(string(output), "AlreadyExists") {
@@ -279,7 +293,7 @@ func smokeTestAllInOneExample(name, yamlFileName string) {
 	jaegerInstance := createJaegerInstanceFromFile(name, yamlFileName)
 	defer undeployJaegerInstance(jaegerInstance)
 
-	err := e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, name, 1, retryInterval, timeout)
+	err := WaitForDeployment(t, fw.KubeClient, namespace, name, 1, retryInterval, timeout)
 	require.NoErrorf(t, err, "Error waiting for %s to deploy", name)
 
 	AllInOneSmokeTest(name)
@@ -292,9 +306,15 @@ func smokeTestProductionExample(name, yamlFileName string) {
 	queryDeploymentName := name + "-query"
 	collectorDeploymentName := name + "-collector"
 
-	err := e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, queryDeploymentName, 1, retryInterval, timeout)
+	if jaegerInstance.Spec.Strategy == "streaming" {
+		ingesterDeploymentName := name + "-ingester"
+		err := WaitForDeployment(t, fw.KubeClient, namespace, ingesterDeploymentName, 1, retryInterval, timeout)
+		require.NoErrorf(t, err, "Error waiting for %s to deploy", ingesterDeploymentName)
+	}
+
+	err := WaitForDeployment(t, fw.KubeClient, namespace, queryDeploymentName, 1, retryInterval, timeout)
 	require.NoErrorf(t, err, "Error waiting for %s to deploy", queryDeploymentName)
-	err = e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, collectorDeploymentName, 1, retryInterval, timeout)
+	err = WaitForDeployment(t, fw.KubeClient, namespace, collectorDeploymentName, 1, retryInterval, timeout)
 	require.NoErrorf(t, err, "Error waiting for %s to deploy", collectorDeploymentName)
 
 	ProductionSmokeTest(name)