+++ test/e2e/wait_util.go
@@ -4,6 +4,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/operator-framework/operator-sdk/pkg/test/e2eutil"
+	"github.com/sirupsen/logrus"
 	"k8s.io/api/extensions/v1beta1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -14,6 +16,7 @@ import (
 // WaitForStatefulset checks to see if a given statefulset has the desired number of replicas available after a specified amount of time
 // See #WaitForDeployment for the full semantics
 func WaitForStatefulset(t *testing.T, kubeclient kubernetes.Interface, namespace, name string, retryInterval, timeout time.Duration) error {
+	start := time.Now()
 	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
 		statefulset, err := kubeclient.AppsV1().StatefulSets(namespace).Get(name, metav1.GetOptions{IncludeUninitialized: true})
 		if err != nil {
@@ -31,15 +34,17 @@ func WaitForStatefulset(t *testing.T, kubeclient kubernetes.Interface, namespace
 		return false, nil
 	})
 	if err != nil {
+		t.Logf("Failed waiting for statefulset after %s\n", time.Since(start))
 		return err
 	}
-	t.Logf("Statefulset available\n")
+	t.Logf("Statefulset available after %s\n", time.Since(start))
 	return nil
 }
 
 // WaitForDaemonSet checks to see if a given daemonset has the desired number of instances available after a specified amount of time
 // See #WaitForDeployment for the full semantics
 func WaitForDaemonSet(t *testing.T, kubeclient kubernetes.Interface, namespace, name string, retryInterval, timeout time.Duration) error {
+	start := time.Now()
 	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
 		daemonset, err := kubeclient.AppsV1().DaemonSets(namespace).Get(name, metav1.GetOptions{IncludeUninitialized: true})
 		if err != nil {
@@ -57,15 +62,17 @@ func WaitForDaemonSet(t *testing.T, kubeclient kubernetes.Interface, namespace, 
 		return false, nil
 	})
 	if err != nil {
+		t.Logf("Failed waiting for daemonset after %s\n", time.Since(start))
 		return err
 	}
-	t.Logf("DaemonSet available\n")
+	t.Logf("DaemonSet available after %s\n", time.Since(start))
 	return nil
 }
 
 // WaitForIngress checks to see if a given ingress' load balancer is ready after a specified amount of time
 // See #WaitForDeployment for the full semantics
 func WaitForIngress(t *testing.T, kubeclient kubernetes.Interface, namespace, name string, retryInterval, timeout time.Duration) (*v1beta1.Ingress, error) {
+	start := time.Now()
 	var ingress *v1beta1.Ingress
 	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
 		ingress, err = kubeclient.ExtensionsV1beta1().Ingresses(namespace).Get(name, metav1.GetOptions{IncludeUninitialized: true})
@@ -84,15 +91,17 @@ func WaitForIngress(t *testing.T, kubeclient kubernetes.Interface, namespace, na
 		return false, nil
 	})
 	if err != nil {
+		t.Logf("Failed waiting for ingress after %s\n", time.Since(start))
 		return ingress, err
 	}
-	t.Logf("Ingress available\n")
+	t.Logf("Ingress available after %s\n", time.Since(start))
 	return ingress, nil
 }
 
 // WaitForJob checks to see if a given job has completed successfuly
 // See #WaitForDeployment for the full semantics
 func WaitForJob(t *testing.T, kubeclient kubernetes.Interface, namespace, name string, retryInterval, timeout time.Duration) error {
+	start := time.Now()
 	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
 		job, err := kubeclient.BatchV1().Jobs(namespace).Get(name, metav1.GetOptions{IncludeUninitialized: true})
 		if err != nil {
@@ -110,15 +119,17 @@ func WaitForJob(t *testing.T, kubeclient kubernetes.Interface, namespace, name s
 		return false, nil
 	})
 	if err != nil {
+		t.Logf("Failed waiting for job after %s\n", time.Since(start))
 		return err
 	}
-	t.Logf("Jobs succeeded\n")
+	t.Logf("Jobs succeeded after %s\n", time.Since(start))
 	return nil
 }
 
 // WaitForJobOfAnOwner checks to see if a given job has completed successfully
 // See #WaitForDeployment for the full semantics
 func WaitForJobOfAnOwner(t *testing.T, kubeclient kubernetes.Interface, namespace, ownerName string, retryInterval, timeout time.Duration) error {
+	start := time.Now()
 	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
 		jobList, err := kubeclient.BatchV1().Jobs(namespace).List(metav1.ListOptions{IncludeUninitialized: true})
 		if err != nil {
@@ -139,15 +150,17 @@ func WaitForJobOfAnOwner(t *testing.T, kubeclient kubernetes.Interface, namespac
 		return false, nil
 	})
 	if err != nil {
+		t.Logf("Failed waiting for job of an owner after %s\n", time.Since(start))
 		return err
 	}
-	t.Logf("Jobs succeeded\n")
+	t.Logf("Jobs succeeded after %s\n", time.Since(start))
 	return nil
 }
 
 // WaitForCronJob checks to see if a given cron job scheduled a job
 // See #WaitForDeployment for the full semantics
 func WaitForCronJob(t *testing.T, kubeclient kubernetes.Interface, namespace, name string, retryInterval, timeout time.Duration) error {
+	start := time.Now()
 	err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
 		cronJob, err := kubeclient.BatchV1beta1().CronJobs(namespace).Get(name, metav1.GetOptions{IncludeUninitialized: true})
 		if err != nil {
@@ -164,8 +177,18 @@ func WaitForCronJob(t *testing.T, kubeclient kubernetes.Interface, namespace, na
 		return false, nil
 	})
 	if err != nil {
+		t.Logf("Failed waiting for cronjob after %s\n", time.Since(start))
 		return err
 	}
-	t.Logf("CronJob succeeded\n")
+	t.Logf("CronJob succeeded after %s\n", time.Since(start))
 	return nil
+}
+
+// WaitForDeployment waits for a deployment to finish and reports how long the operation took
+func WaitForDeployment(t *testing.T, kubeclient kubernetes.Interface, namespace, name string, replicas int, retryInterval, timeout time.Duration) error {
+	start := time.Now()
+	err := e2eutil.WaitForDeployment(t, fw.KubeClient, namespace, name, 1, retryInterval, timeout)
+	elapsed := time.Since(start)
+	logrus.Infof("Deployment of %s in namespace %s took %s\n", name, namespace, elapsed)
+	return err
 }